I've established menus and drop-down menus as well as drop-down menus that have menus inside of their menus so this is pretty coolgit status

+++ FunGUI/src/fungui/IncrementalSlider.java
@@ -19,7 +19,7 @@ public class IncrementalSlider extends Slider {
 	
 	@Override
 	public float percent() {
-		return(increment() / Math.round(range()));
+		return((float) (increment()) / Math.round(range()));
 	}
 	
 	public int increment() {
